+++ SFML_RPG/EnemySystem.h
@@ -20,7 +20,7 @@ private:
 	
 	const TileMap& tileMap;
 	float despawnRange = 1500.f;
-	float totalDamage = 0.f;
+	int totalDamage = 0;
 
 	float trackPlayer(Enemy* enemy, sf::Vector2f& playerPosition, const float& dt) const;
 
@@ -29,7 +29,7 @@ public:
 	virtual ~EnemySystem();
 
 	const std::vector<Enemy*> getEnemies() const;
-	float getTotalDamage();
+	int getTotalDamage();
 	void loadSpawners();
 
 	void createEnemy(Type type, EnemySpawner& spawner);
